+++ tests/testthat/test_calculateFitnessScore.R
@@ -35,7 +35,8 @@ Ldef <- list(
     filteredReadsFastqReverse = "",
     maxNReads = -1, 
     verbose = FALSE,
-    nThreads = 1, chunkSize = 1000
+    nThreads = 1, chunkSize = 1000, 
+    maxReadLength = 1024
 )
 Ldef1 <- c(
     list(fastqForward = system.file("extdata/cisInput_1.fastq.gz", package = "mutscan"),